Refactorize NN Components: Setup -> Runner
